Update License (#29)
Summary of changes:

All contributions will be distributed under the GNU Affero General Public License (AGPL) v3. This ensures that the source code remains open and accessible, even when used over a network.

The right to offer contributions under more permissive licenses to commercial companies is also reserved. This will be used if they contribute back to the project positively (financially or otherwise). This allows us to sustain and grow the project without taking away the AGPL option.